feat: New API
