+++ src/data/index.js
@@ -1,0 +1,5 @@
+'use strict'
+
+module.exports = {
+  json: require('./json')
+}
