+++ test/node.js
@@ -1,3 +1,3 @@
 'use strict'
 
-require('./json')
+require('./data/json')
